+++ gfs_physics.c
@@ -1,14 +1,13 @@
 /*
- * FILE      Code\Physics.cpp
+ * FILE      gfs_physics.c
  * AUTHOR    Ilya Akkuzin <gr3yknigh1@gmail.com>
  * COPYRIGHT (c) 2024 Ilya Akkuzin
  * */
-#include "Physics.hpp"
-
-#include "Types.hpp"
+#include "gfs_physics.h"
+#include "gfs_types.h"
 
 bool
-RectangleU16IsInside(RectangleU16 r, u32 x, u32 y) {
+RectangleU16IsInside(RectangleU16 r, u16 x, u16 y) {
     return x >= r.x && x <= r.x + r.width && y >= r.y && y <= r.y + r.height;
 }
